Remove unused typing imports in spec_flow
Summary:
## Instructions about RACER Diffs:
You will receive **FULL CREDIT** (ETS) for this diff if you either:
 - [**Recommended**]Commandeer and land this diff

Once landed, feel free to claim the associated task and the EYS project.

 This diff is removing the unused imports from Code Quality Score from [Monetization codehub](https://fburl.com/quality/58ozwacb):
 - Accept and ship the diff (racer do not do auto land)
 - [**Recommended**]Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generage.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T229209311. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [RACER Feedback and Q&A](https://fb.workplace.com/groups/2477474979269093) group.
## Summary:
Removed unused imports `Dict` and `Set` from the typing module in spec_flow.py as they were not being used anywhere in the file. This addresses the flake8 F401 issue identified in T229209311.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=6958aa9e-54ad-11f0-a495-fa784727a89a&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=6958aa9e-54ad-11f0-a495-fa784727a89a&tab=Trace)

Differential Revision: D77502033

fbshipit-source-id: d6011248923cfc66f9cb1df4f53452d48eae0a5a